2.12.8 - When current action is Clarify & Estimate, show an input field with label Dev Time Estimation in Task Detail tab

+++ database/migrations/2024_08_14_100100_create_tickets_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('name');
             $table->longText('composed_name')->nullable();
             $table->integer('type');
-            $table->bigInteger('initial_estimation_development_time');
+            $table->float('initial_estimation_development_time', 8, 2);
             $table->longText('release_note')->nullable();
             $table->boolean('auto_wait_for_client_approval')->default(0);
             $table->integer('status')->nullable();
